Add adjusment

+++ src/FSI.AccessAuthentication.Domain/Entities/AuthenticationEntity.cs
@@ -2,7 +2,7 @@
 {
     public class AuthenticationEntity : BaseEntity
     {
-        public int SystemId { get; set; }
+        public long SystemId { get; set; }
         public string Username { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
         public bool IsAuthentication { get; set; }
